+++ packages/formik/src/Field.tsx
@@ -1,0 +1,16 @@
+import * as React from 'react';
+import { forwardRefWithAs } from './utils/hooks';
+import { useField } from './useField';
+
+export type FieldConfig = React.HTMLProps<HTMLInputElement> & { name: string };
+
+export const Field = forwardRefWithAs<FieldConfig, 'input'>(
+  React.memo(function Form({ as: Comp = 'input', ...props }, forwardedRef) {
+    const [field] = useField<any>({ ...props, as: Comp });
+    return <Comp ref={forwardedRef} {...field} {...props} />;
+  })
+);
+
+if (__DEV__) {
+  Field.displayName = 'Field';
+}
