+++ packages/formik/src/utils/canUseDOM.ts
@@ -1,0 +1,7 @@
+export function canUseDOM() {
+  return (
+    typeof window !== 'undefined' &&
+    typeof window.document !== 'undefined' &&
+    typeof window.document.createElement !== 'undefined'
+  );
+}
